Переименовал файлы и разделил функции по файлам в соответствии с типом используемых алгоритмов. Перенёс в файл system_functions.py служебные функции (загрузка и выгрузка данных). Начал работу над добавлением дополнительного аргумента командной строки.
